# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Premium Car Dealership API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /cars:
    get:
      operationId: listCars
      tags: [cars]
      summary: List cars with filters
      parameters:
        - name: make
          in: query
          schema: { type: string }
        - name: model
          in: query
          schema: { type: string }
        - name: yearMin
          in: query
          schema: { type: integer }
        - name: yearMax
          in: query
          schema: { type: integer }
        - name: priceMin
          in: query
          schema: { type: number }
        - name: priceMax
          in: query
          schema: { type: number }
        - name: bodyType
          in: query
          schema: { type: string }
        - name: fuelType
          in: query
          schema: { type: string }
        - name: transmission
          in: query
          schema: { type: string }
        - name: condition
          in: query
          schema: { type: string }
        - name: color
          in: query
          schema: { type: string }
        - name: availability
          in: query
          schema: { type: string }
        - name: featured
          in: query
          schema: { type: boolean }
        - name: category
          in: query
          schema: { type: string }
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, oldest, price_asc, price_desc, mileage_asc, featured]
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
        - name: search
          in: query
          schema: { type: string }
      responses:
        "200":
          description: List of cars
          content:
            application/json:
              schema:
                type: object
                properties:
                  cars:
                    type: array
                    items:
                      $ref: "#/components/schemas/Car"
                  total:
                    type: integer
                  offset:
                    type: integer
                  limit:
                    type: integer
                required: [cars, total, offset, limit]
    post:
      operationId: createCar
      tags: [cars]
      summary: Create a new car listing
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarBody"
      responses:
        "201":
          description: Created car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
  /cars/featured:
    get:
      operationId: getFeaturedCars
      tags: [cars]
      summary: Get featured cars for homepage
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 6 }
      responses:
        "200":
          description: Featured cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"
...
  /cars/{id}:
    get:
      operationId: getCar
      tags: [cars]
      summary: Get a single car by ID
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Car detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
        "404":
          description: Not found
    put:
      operationId: updateCar
      tags: [cars]
      summary: Update a car
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateCarBody"
      responses:
        "200":
          description: Updated car
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
    delete:
      operationId: deleteCar
      tags: [cars]
      summary: Delete a car
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "204":
          description: Deleted
  /cars/{id}/slug:
    get:
      operationId: getCarBySlug
      tags: [cars]
      summary: Get car by slug
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: Car detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Car"
  /cars/{id}/related:
    get:
      operationId: getRelatedCars
      tags: [cars]
      summary: Get related cars
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
        - name: limit
          in: query
          schema: { type: integer, default: 4 }
      responses:
        "200":
          description: Related cars
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Car"

  # ─── INQUIRIES ─────────────────────────────────────────────────────────────
  /inquiries:
    get:
      operationId: listInquiries
      tags: [inquiries]
      summary: List all inquiries
      parameters:
        - name: status
          in: query
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Inquiries list
          content:
            application/json:
              schema:
                type: object
                properties:
                  inquiries:
                    type: array
                    items:
                      $ref: "#/components/schemas/Inquiry"
                  total:
                    type: integer
                required: [inquiries, total]
    post:
      operationId: createInquiry
      tags: [inquiries]
      summary: Submit a customer inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateInquiryBody"
      responses:
        "201":
          description: Created inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
  /inquiries/{id}:
    get:
      operationId: getInquiry
      tags: [inquiries]
      summary: Get a single inquiry
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      responses:
        "200":
          description: Inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"
    put:
      operationId: updateInquiry
      tags: [inquiries]
      summary: Update inquiry status
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: integer }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                notes:
                  type: string
              required: [status]
      responses:
        "200":
          description: Updated inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Inquiry"

  # ─── BOOKINGS ──────────────────────────────────────────────────────────────
...
  /financing:
    get:
      operationId: listFinancingInquiries
      tags: [financing]
      summary: List financing inquiries
      parameters:
        - name: limit
          in: query
          schema: { type: integer, default: 20 }
        - name: offset
          in: query
          schema: { type: integer, default: 0 }
      responses:
        "200":
          description: Financing inquiries
          content:
            application/json:
              schema:
                type: object
                properties:
                  inquiries:
                    type: array
                    items:
                      $ref: "#/components/schemas/FinancingInquiry"
                  total:
                    type: integer
                required: [inquiries, total]
    post:
      operationId: createFinancingInquiry
      tags: [financing]
      summary: Submit a financing inquiry
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateFinancingInquiryBody"
      responses:
        "201":
          description: Created financing inquiry
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/FinancingInquiry"
components:
  schemas:
...
    Car:
      type: object
      properties:
        id:
          type: integer
        slug:
          type: string
        title:
          type: string
        make:
          type: string
        model:
          type: string
        trim:
          type: string
          nullable: true
        year:
          type: integer
        price:
          type: number
        discountedPrice:
          type: number
          nullable: true
        mileage:
          type: integer
        engineSize:
          type: string
          nullable: true
        transmission:
          type: string
        fuelType:
          type: string
        drivetrain:
          type: string
          nullable: true
        bodyType:
          type: string
        seats:
          type: integer
          nullable: true
        doors:
          type: integer
          nullable: true
        color:
          type: string
        stockNumber:
          type: string
          nullable: true
        vin:
          type: string
          nullable: true
        condition:
          type: string
        availability:
          type: string
        shortDescription:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
          nullable: true
        safetyFeatures:
          type: array
          items:
            type: string
          nullable: true
        comfortFeatures:
          type: array
          items:
            type: string
          nullable: true
        techFeatures:
          type: array
          items:
            type: string
          nullable: true
        gallery:
          type: array
          items:
            type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPublished:
          type: boolean
        financingAvailable:
          type: boolean
        location:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - slug
        - title
        - make
        - model
        - year
        - price
        - mileage
        - transmission
        - fuelType
        - bodyType
        - color
        - condition
        - availability
        - isFeatured
        - isPublished
        - financingAvailable
        - createdAt
        - updatedAt
    CreateCarBody:
      type: object
      properties:
        title:
          type: string
        make:
          type: string
        model:
          type: string
        trim:
          type: string
          nullable: true
        year:
          type: integer
        price:
          type: number
        discountedPrice:
          type: number
          nullable: true
        mileage:
          type: integer
        engineSize:
          type: string
          nullable: true
        transmission:
          type: string
        fuelType:
          type: string
        drivetrain:
          type: string
          nullable: true
        bodyType:
          type: string
        seats:
          type: integer
          nullable: true
        doors:
          type: integer
          nullable: true
        color:
          type: string
        stockNumber:
          type: string
          nullable: true
        vin:
          type: string
          nullable: true
        condition:
          type: string
        availability:
          type: string
        shortDescription:
          type: string
          nullable: true
        description:
          type: string
          nullable: true
        features:
          type: array
          items:
            type: string
          nullable: true
        safetyFeatures:
          type: array
          items:
            type: string
          nullable: true
        comfortFeatures:
          type: array
          items:
            type: string
          nullable: true
        techFeatures:
          type: array
          items:
            type: string
          nullable: true
        gallery:
          type: array
          items:
            type: string
          nullable: true
        videoUrl:
          type: string
          nullable: true
        isFeatured:
          type: boolean
        isPublished:
          type: boolean
        financingAvailable:
          type: boolean
        location:
          type: string
          nullable: true
        category:
          type: string
          nullable: true
      required:
        - title
        - make
        - model
        - year
        - price
        - mileage
        - transmission
        - fuelType
        - bodyType
        - color
        - condition
        - availability
...
    Inquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        message:
          type: string
        type:
          type: string
        status:
          type: string
        notes:
          type: string
          nullable: true
        createdAt:
          type: string
      required: [id, name, phone, message, type, status, createdAt]
    CreateInquiryBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        message:
          type: string
        type:
          type: string
      required: [name, phone, message, type]
...
    FinancingInquiry:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        loanAmount:
          type: number
          nullable: true
        downPayment:
          type: number
          nullable: true
        loanTermMonths:
          type: integer
          nullable: true
        employmentStatus:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
        status:
          type: string
        createdAt:
          type: string
      required: [id, name, phone, status, createdAt]
    CreateFinancingInquiryBody:
      type: object
      properties:
        name:
          type: string
        phone:
          type: string
        email:
          type: string
          nullable: true
        carId:
          type: integer
          nullable: true
        carTitle:
          type: string
          nullable: true
        loanAmount:
          type: number
          nullable: true
        downPayment:
          type: number
          nullable: true
        loanTermMonths:
          type: integer
          nullable: true
        employmentStatus:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      required: [name, phone]