# lib/api-zod/src/generated/types/financingInquiry.ts
/**
 * Generated by orval v8.5.3 🍺
 * Do not edit manually.
 * Api
 * Premium Car Dealership API
 * OpenAPI spec version: 0.1.0
 */

export interface FinancingInquiry {
  id: number;
  name: string;
  phone: string;
  email?: string | null;
  carId?: number | null;
  carTitle?: string | null;
  loanAmount?: number | null;
  downPayment?: number | null;
  loanTermMonths?: number | null;
  employmentStatus?: string | null;
  message?: string | null;
  status: string;
  createdAt: string;
}